Updated to use fake-auth.netlify.com
